+++ CHANGES.rst
@@ -3,8 +3,8 @@ Changelog
 
 
 v2.3.1 (2014-12-07)
--------------------
 
+- Add ``pipe`` and ``pipe_right`` as aliases of ``flow`` and ``flow_right``.
 - Fix ``merge`` so that trailing ``{}`` or ``[]`` don't overwrite previous source values.
 - Make ``py_`` an alias for ``_``.